requirements.txt
Django>=5.0,<6.0
gunicorn>=21.2
requests>=2.31
python-dotenv>=1.0
cryptography>=42.0
whitenoise>=6.6
